requirements.txt
pyyaml
pydantic
argparse
rich
